NRL-2099 List apps for consumer & producer perms
